Added entityType to paginator
BREAKING CHANGE: signature for `getPaginator` includes a new required parameter.

This resolves an issue where, if one has multiple paginators, there is a possibility
that the id space for each will end up being corrupted as multiple reducers will be
processed for each action. This is especially problematic if there is overlap in the
ids for entities or if there are custom entityIdKeys on the responses from the API...
so basically any time there's more than one paginator.